Optimized: JSON-RPC playground with multiple options
In order to test other functionalities, the JSON-RPC playground (for development only) has now a dropdown selector which allows to switch between other playgrounds which are going to be implemented.
